feat: move signature vector to data uri to shared
