+++ package.json
@@ -77,7 +77,6 @@
     "boxicons": "1.8.0",
     "bson": "^5.5.0",
     "busboy": "^1.6.0",
-    "canvas": "^3.1.2",
     "celebrate": "^15.0.3",
     "compression": "~1.8.1",
     "connect-datadog": "0.0.9",
@@ -132,7 +131,6 @@
     "openid-client": "^6.6.2",
     "openid-client-legacy": "npm:openid-client@^5.7.1",
     "opossum": "^8.1.4",
-    "perfect-freehand": "^1.2.2",
     "promise-retry": "^2.0.1",
     "promise-timeout": "^1.3.0",
     "puppeteer-core": "24.9.0",
