+++ shared/package.json
@@ -3,11 +3,13 @@
   "version": "1.0.0",
   "description": "",
   "dependencies": {
+    "canvas": "^3.2.0",
     "date-fns": "^3.6.0",
     "json-stringify-safe": "^5.0.1",
     "jszip": "^3.7.1",
     "libphonenumber-js": "^1.10.48",
     "lodash": "^4.17.21",
+    "perfect-freehand": "^1.2.2",
     "tweetnacl": "^1.0.3",
     "tweetnacl-util": "^0.15.1",
     "type-fest": "^4.17.0",
